// src/models/slideInterface.ts
export interface Slide {
    id?: string
    offerID: string
    url: string
    href?: string
    order: number
    alt: string
    // href?: string
    // clickAction?: Function
    // caption?: string
    // title?: string
    // backgroundSize?: string
    // backgroundPosition?: string
    // backgroundRepeat?: string
}